Updated README.md
Grammar changes, clarification of one of the attached demos (gif is of classification, so the demo section is about that. I added a sentence stating that the classification can be used for sentiment analysis). Also changed the runserver command to utilize port 8080, as that is the one specified in the localhost URL.